requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cryptography==43.0.1
cssselect2==0.7.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
fonttools==4.53.1
html5lib==1.1
inflection==0.5.1
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
PyMySQL==1.1.1
pyphen==0.16.0
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
uritemplate==4.1.1
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.7.0
zopfli==0.2.3
